Work around https://github.com/docker/docker/issues/11087
This is a rather simple work around that conviently worked out.  The
python-agent already has code in it such that if its parent process
dies it will terminate itself.  This logic was added to prevent creating
orphan agents in the SSH agent scenario (which is no longer used in
rancher).  By changing the bootstrap script to exec the python agent and
not fork it as a child, this causes the python-agent's parent to be the
container PID.  When the container PID is killed, the python-agent still
has open the FDs from the container PID.  This prevents the container
from properly dying.  But since the python-agent will kill itself when
the container PID dies, the entire container will cleanup.