Native docker integration
Support the discovery of containers created outside of Rancher.

The goal of this commit is to create Rancher containers to represent
real docker containers created on Rancher hosts outside of the Rancher
API. To accomplish this, we monitor docker events and push them up to
Rancher and Rancher then creates container resources that are utlimately
synced up with the corresponding native docker containers.

This commit supports the following:
* Native container creation
* Monitoring when a native container is stopped (via the pre-existing
ping monitor logic)
* Performing Rancher actions such as start, stop, and restart on native
containers.
* Being able to specify that native container should get Rancher
networking.

This commit does not support:
* Discovering when containers (native and Rancher) are started/restarted
outside of Rancher via events
* Fully populating the Rancher container with all the data available
from the docker inspect
* Full syncing/importing of pre-existing docker containers