Support sync and import
Sync up all containers in ping monitor. The previous implmentation only
synced stopped containers that rancher already new about. The new
implementation will attempt to sync all containers regardless of their
state and whether Rancher previously knew about them. This allows us to
import containers that existed on a host prior to starting
rancher-agent on it.

To allow for proper syncing of purged containers, also added an agent
based process handler for `instancehostmap.remove` so
that when that process is executed by InstancePurge, a
`compute.instance.remove` event is sent to python-agent, which will
cause the container to be removed. Prior to this change, the container was
not removed until its corresponding root volume was purged, which happened
independently and a significant amount of time after instance.purge
successfully completed.

Also, when a docker restart occurs, the start and stop events come in so
quickly that the start doesn't get to execute because stopping is not a
valid state for the start process. In this scenario, wait for stopping
to complete first.