Include port number in haproxy config for hdr(host) acl.
In order for HA proxy to route traffic based on the Host header, when the client includes the port number in the Host header, the acl must also include the port number in order for it to match.

See http://serverfault.com/questions/502443/ignore-port-numbers-in-haproxy-host-header-matches for a related issue that describes the use case this solves.