Don't manage ports for non-rancher networks
Only configure ports on the host for containers that are part of a
Rancher managed network (and more explicitly, one that has the
portService network service).

This change is prompted by a bug where a rancher/server container
becomes unavailable on 8080 if it a host that also has a rancher/agent
running.

After looking at the bug, we decided that if the user chooses a
non-rancher network, Rancher should not be doing any port manipulation
at all.

Bonus change: Delete missing containers

Change behavior of ping to be more consistent and logic such that if a
container is deleted on the host when rancher-agent is down, that
container gets marked as deleted, as opposed to stopped (which is the
current behavoir).