Lookup named volumes on instance.create
Adds logic to support properly scheduling containers that are using
shared volumes. If a named volume is specified in dataVolumes like
this: `vol1:/container/path`, then during instance.create, look to see
if that volume exists in a non-local storage pool and if it does, use
that volume by adding an entry to dataVolumeMounts.