Fix race condition for volume external event
If an external volume event comes in at the same time as we are doing
native-docker population, it is possible to hit a race condition where
a volume is created twice. This fix prevents that by doing an
in-transaction lookups of the volume in the storage pool before creating
it.