Don't look for instances on every agent ping
In the past we have had issues in which `docker ps` is not as fast as it
should be.  This causes agents to disconnect because they get hung on
`docker ps`.  By decreasing the amount of times we call `docker ps` will
allow some pings to go through which should hopefully allow the agents
to stay connected.