Apply config items to network agents but don't wait
When a network agent starts it changes important networking information
should be reflected in the environment.  Previously a network agent start
would update no other network agents.  This caused a race condition in
which other network agents may or may not know about the newly launched
host.

With this change we will now apply the config items but we will not wait
for them to complete.  We obviously can't wait for completion because we
are in the process of starting the network agent that is the one that must
complete the action.