Allow labels on host create only
When a host is created the labels should be assigned that were passed in
then CATTLE_HOST_LABELS environment variable.  If the user changes the
label after the fact, we should use the value the user provided and not
the value from CATTLE_HOST_LABELS.  This is accomplished by only using
CATTLE_HOST_LABELS when creating the host entry but not updating it.