don't process init health check event
as its being set by cattle on instance restart.
That is done to avoid the scenario when init state is reported on
healthcheck process restart inside the agent happening around the time
when instance becomes unheatlhy. This can lead to instance being set
with reinitializing state instead of unheatlhy, and it postpones (or
even cancels, if reinitializing timeout is not set) instance recreation