Dont delete shared pools when an agent is removed
If a storage pool agent container goes unhealthy, it will get removed
and replaced. This causes the associated agent to get removed and
replaced. This change prevents the storage pool assocaited with the
agent from being removed so that when the new storage pool agent
container comes online, it can be associated with the existing pool.