Convert event exceptions in FailedToAllocate
When the scheduler replies with an error, that was being bubbled up as
an EventExecutionException. This type of exception prevented the
instance from ever failing to allocated, it just got stuck in starting.

This change catches those exceptions, logs their message (which
contains the error message sent by the scheduler) and rethrows
FailedToAllocate exceptions, which will cause the instance to go to an
error state.