Allow host resource fields on create
This lets the user set the milliCpu, localStorageMb, and memory fields
when creating a host via our machine integration.

The current agent backpopulate logic copies ALL fields over on the
first sync. So, this change makes sure the relevant fields are removed
from the update map if they are already set on the host.