Set resourceId on scheduling events
Without it set, the event framework in the scheduler was locking on ""
and thus all events were locking on the same key and causing events to
drop.

Changed how resouceType because it impacts the actual id in the event
because it is transformed into obfuscated form
(i<resourceType abbreviation><id>).