Fix volume locking issue
Fixes a bug where two volume storage pool maps were getting created
when two instances that used the volume where scheduled at near the
same time.

Also fixes an issue where allocation would timeout because
AllocationConstrainLock was not a blocking lock, which caused locking
to fail and not be retried for 15 seconds.