Provide access and secret keys when getting client
This fin() method started erroring out because when no keys are found
in the environment variables, a tuple with two None values are passed
to the requests library.

I don't know how this ever worked. My best guess is that up until now,
CATTLE_ACCESS_KEY and CATTLE_SECRET_KEY were being found as env vars
and being used.

This change is actually a better test of auth being off because we
don't want previously valid keys to be used when affirming that auth
has been switched off.