Remove conflicting setVersion and unnecessary check
In the filter delegate, the block of code right before the line being
removed is concerned with setting version, so it makes no sense to
unset version a few lines later.

In the whitelisted proxy, since version will always be set, it makes no
sense to check that it is not set.