Issue tokens yearly
Each registration token we issue is good for 10 years.  The problem is
that we issue a new one each minute and people assume that means the
last one isn't valid.  So now we will just issue a new one each year,
which is still valid for 10 years.  You can always invalidate the tokens
by deleting the registationToken in the UI.