Re-enable public endpoints test
This test was disabled while moving port allocation logic to the
external scheduler. Moving that logic to the external scheduler made
this test fail because it relied on the ports to force instances of the
same service to different hosts. With the external scheduler disabled
in the integration tests, that was not happening.

The fix was to force the instances of the service to different
hosts by using the global scheduling rule instead of just setting
scale=2.

We'll need an additional test to assert that "0.0.0.0" behaves
the same as having no IP address, but that will be committed separately
when that issue is fixed.