Fail svc create if out of environment ports
When a service does not specify a public port in its port map, cattle
will allocate an environment scoped port to it. Then every container
created for that service will get the same port exposed on the host.

Environment ports have a defined range and the range can be exhuasted.
The previous behavior when the range was exhausted was to simply allow
the service to be created with undefined behavior for what would happen
to the ports. With the new port allocation logic introduced by migrating
that logic to the external scheduler, the service containers were
getting a randomly assigned port. This is the correct behavior if a
container doesn't have a port, but it doesn't make sense for the
environment scoped ports of services, where consistency across
containers in a port is necessary.

This fixes the problem by failing to create the service if the port
range is exhausted. The service will remain in registering but will
have a transitioning message indicating the error.