Fix scheduler spreading
This fix addresses a bug where containers were not spreading at all
when the external scheduler was enabled. All containers would go to the
first host that the scheduler knew about.

The cause of this bug was that we were not sending the "phase" to the
external scheduler for the prioritize event. The prioritize event is
sent by cattle to ask for a prioritized list of hosts where the fist
item in the list is the best fit.

We added the phase logic to accomodate releasing ports on instance.stop.
The scheduler makes a decision on what resources to prioritize,
allocate, and deallocate based on the phase. If the phase is
instance.allocate or intance.deallocate, it will consider all possible
resources (container count, cpu, ports). If the phase is
instance.start|stop, it will just consider ports.

For the fix, I've refactor the logic to guarantee that a phase is
always required and set in a single location.