Fix 'secrets' volume allocation
In a multi-host environment, deploying a container with a secret would
only succeed on a single host. This is because the allocation logic for
volumes was mishandling the scenario where a storage driver had more
than one storage pool associated with it. This is the case for secrets
because each host gets its own pool.

The problem was that the constraint would only allow the first pool in
the list of pools associated with the storage driver to be used. The
fix was to allow all pools assocaited with the driver to be used.