Support legacy port allocation
Support the scenario where the user has upgraded rancher but not
upgraded the scheduler service. To do this, I've reintroduced the
previous port constraint logic and added a check to see if the
scheduler is of a version that is less than the version in which we
added port allocation logic.