Add host event create for scheduler update
We need to be able to ensure that when a host is added, removed, or
updated, global service reconciliation is kicked off AFTER the external
scheduler has had a chance to update its internal state.

This event handler allows for that. The flow for CRUDing the host will
be thus:
1. Host is CRUDed in cattle
2. Existing service reconcile trigger logic runs, but the host change
has not yet propogated to metadata and thus it is not in the scheduler,
so the scheduler behaves as though it doesnt know about the change and
global services are not changed.
3. Metadata updates and scheduler updates its internal state
4. Scheduler sends a external host event informing cattle that it
updated its state and global service reconciliation needs to be
triggered.
5. ExternalHostSchedulerUpdateEventCreate handles the event and kicks
off reconcile for global services.