Don't call external sched during deploy planning
The global service deployment planner was calling the external
scheduler to check if any hosts should be excluded from its list due to
scheduling rules there. This was specifically added to allow the
io.rancher.scheduler.require_any host label to work with global
services. But we cannot make this call here because the calling code
occupies a thread in the core executore service, which must be fast or
it will back up the system.