Reconcile healthchecks on host.activate
This fixes a bug that occurs if you take you environment down to zero
hosts by removin the hosts from your cloud provider and they go into a
disconnected state in Rancher. In that scenario, healthchecked
instances on the disconnected host would never get healthchecks
scheduled on the newly added hosts and thus would never get moved off
the disconnected host.

The process that we were previously keying off of
(networkserviceproviderinstancemap.create) is no longer executed. It
was executed to lanuch the network-agent container on the host, but
we've completely gone away from that model. We were previously using
that process because we needed the network-agent to be running in order
for that host to process healthchecks, but that is no longer necessary
because healthcheks are deployed in their own containers. The logic
later in this process that schedules the new healthchecks ensures that
the host will have a healthcheck container deployed to it, if it
doesn't already.