Handle duplicate instance host maps gracefully
When stopping or starting an instance and calling the external
scheduler, we have to get the host UUID to send to the scheduler. In
the very odd scenario that more than one instance host map exists,
first determine if they are all pointing at the same host and if so,
return that host. If they do not all point at the same host, just log
an error and return null so that stop/start isn't stuck forever.