Dont find an agent if the resource is removed
When getting the "agentResource" that is later used to find an agent,
don't return the resource if it has been removed.

This will cause AgentBasedProcessHandler logic to short circuit and
complete successfully.

The concrete case for this if when a storage pool has been removed but
there are still volumeStoragePoolMaps associated with it that need
removed.