add tweaks to recognize special bind mounts
We are rewriting the volume logic for rancher managed storage driver. The rancher managed volume are passed as a bind mount to the container. So we need to parse the mountpoint to see if its a rancher mountpoint.
If so, we make this mountpoint tied to the specific volume so that API/UI would know the active mounts for the volume and show the correct state.