Set driver and device number on stack vols
These two fields need to be set properly in order for the agent to know
how to handle volume removal. We enforce that driver is not empty via
API validation and in the other places in code where we create volumes,
but just missed this one. Same with device number.