Dont fallback to deactivate+remove for stack vols
When deleting volumes owned by a stack, if we get a process cancel
exception on the volume.remove, we do a deactivate chained to remove.
This is because the containers using the stack scoped volume may not
have been deleted yet and thus it is not right to deactivate the volume.

This change will let the ProcessCancelException bubble up which will
caused the stack remove to be rescheduled for some seconds later. At
that time, it will try again and if the volume has been deactivated by
the normal lifecycle of removing the contaoiner(s) that was using it,
it will be removed.