Use event for instance data cache invalidation
This change adresses a bug where in an HA setup, when a load balancer
was updated to have a new port rule, the newly published port may not
show up in metadata and thus not get programmed properly by networking.
The bug was caused by not invalidating the instance data cache on all
nodes and thus potentially serving up stale data to the metadata
service.