Restrict access to infrastructure
Disable machine config link using setting - Old way was to guess
if machine config link should be there based on whether or not
the user had access to create hosts. This makes it more explicit
and groups it with other "restricted" functionality.

Limit host actions based on infrastructure setting - Only allow
roles whitelisted in the modify.infrastructure.roles setting to
perform actions on hosts and physicalhosts or to access a machine's
config.

Restrict access to registrationTokens - Only allow roles whitelisted
in the modify.infrastructure.roles to access registrationTokens.