Handle http response from auth-service
returnResponse() method that was being used does not dispose off connections properly.
AuthServiceLauncher was not using discardContent after calling reload on auth-service.
This led to connection pool getting exhausted on the http requests and the rancher server froze.
This commit uses handleResponse() and discardContent() to handle the response and prevent connection pool from getting exhausted.