Report resource constraints when scheduling fails
We previously implemented a change to include resource constains in the
error message presented to users when the external scheduler could not
find any matching hosts. However, we missed a case where the external
scheduler found an inactive host that was later filtered out by cattle
due to the fact that it was inactive.

This fix addresses that change by moving the logic for including the
resource constraint up a level and always including the constaints if
they are present.