Ignore errors during catalog refresh on startup
A change in the catalog-service refresh api behavior necessiates this
change so that cattle can move forward with starting when
catalog-service fails to refresh catalogs.

The api behavior has changed to return a 500 response code f a refresh fails.
Previously, it was just swallowing the errors and always returning a 200.
This, combined with the fact that it would take 10 minutes to return
a response in airgap setups, meant that cattle would effectively block
for 10 minutes in airgap setups before creating any project
templates/infrastructure stacks.