be able to sync k8s containers
the k8s container are launched with network mode: container and is pointing to its sandbox. But we don't back-populate the pause container in rancher so this network container look up could happen forever. So remove the check if it is k8s pod. The side-effect would be if we launched a native pod in k8s with container networking, this is possible that we will most likely to drop the container field in back-populated containers. 😐