Fix Hazelcast message amplification
Cattle will create topics such as "ping;agent=4" on demand per agent.  The
approach taken previously was to create one topic named "ping;" and then
multiplex all "ping;*" messages over the same topic.  This kept down the
number of topics we create, but also avoided the need to have some type
of GC logic of deleting topics.

If two topics, "ping;agent=1" and "ping;agent=2" were needed then one
topic "ping;" was created with two subscriptions.  The subscribers would
then check the message key if it matched "agent=1" or "agent=2".

Unfortunately this approach has some fairly bad side effects in
Hazelcast which has lead to a fairly large amplification of traffic.
What has been often seen is if you X bytes flowing over port 8080, you'd
have 10X flowing over port 9345.

The reason for this is that when one publishes to a topic there is one
message sent per subscriber, not one per node that has a subscriber.
The later was assumed to be the behavior.  For example, if node A
published a message to a topic that has 10 subscribers on node B, node B
will get 10 messages across the wire.  Not one message for node B that
is then seen by the 10 subscribers.

Topics are used quite a bit in Cattle but one distinct example is
pinging nodes.  One topic per node (which is described as an agent in
the system) is created.  If one had 100 nodes, in the previous code, 1
topic is created with 100 subscriptions.  Every 15 seconds an agent is
pinged.  100 nodes would translate to 100 published messages every 15
seconds.  Given an HA setup with worst case distribution you would
expect 100 messages across the wire every 15 seconds. Unfortunately,
due to how Hazelcast works it is actually N^2 every 15 seconds, so 10000
messages.

To fix this issue we no longer multiplex messages across topics and
create one topic per full name, thus reducing the subscription count.
This does mean we dynamically create and delete topics now.  GC logic
has been added such that a topic with 0 subscription for 15 seconds will
be deleted.  If that topic is needed later it will be created again.